# Join ensembles more efficiently