This PR brings improvements to first-run compilation time.
## Benchmark
main:
```
julia> using GeoStatsTransforms, GeoStatsProcesses, GeoStatsFunctions, GeoTables, Meshes

julia> gtb = georef((; a=rand(10, 10)));

julia> dom = CartesianGrid(100, 100);

julia> @time gtb |> Simulate(dom, 100, :a => GaussianProcess(GaussianVariogram()));
Simulating a: 100%|████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:15
 52.954089 seconds (160.39 M allocations: 18.220 GiB, 5.65% gc time, 77.95% compilation time: <1% of which was recompilation)

julia> @time gtb |> Simulate(dom, 100, :a => GaussianProcess(GaussianVariogram()));
Simulating a: 100%|████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:10
 10.620602 seconds (109.13 M allocations: 14.898 GiB, 11.41% gc time)
```
PR:
```
julia> using GeoStatsTransforms, GeoStatsProcesses, GeoStatsFunctions, GeoTables, Meshes

julia> gtb = georef((; a=rand(10, 10)));

julia> dom = CartesianGrid(100, 100);

julia> @time gtb |> Simulate(dom, 100, :a => GaussianProcess(GaussianVariogram()));
Simulating a: 100%|████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:14
 21.125534 seconds (131.76 M allocations: 16.378 GiB, 11.83% gc time, 46.31% compilation time: <1% of which was recompilation)

julia> @time gtb |> Simulate(dom, 100, :a => GaussianProcess(GaussianVariogram()));
Simulating a: 100%|████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:10
 10.823119 seconds (109.12 M allocations: 14.896 GiB, 9.55% gc time)
```